requirements.txt
# Combined requirements for API + Worker + LLM_base
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
redis[asyncio]>=5.0
requests>=2.31.0
PyYAML>=6.0
aiosqlite>=0.22
